src/requirements.txt

free-proxy==1.1.1
proxy-checking==0.0
python-telegram-bot==20.3

